--- a/packages/turf-clean-coords/tsconfig.testTypes.json
+++ b/packages/turf-clean-coords/tsconfig.testTypes.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.shared.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["types.ts"]
+}