--- a/packages/turf-clusters-dbscan/tsconfig.testTypes.json
+++ b/packages/turf-clusters-dbscan/tsconfig.testTypes.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.shared.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["types.ts"]
+}