--- a/packages/turf-clusters-kmeans/tsconfig.testTypes.json
+++ b/packages/turf-clusters-kmeans/tsconfig.testTypes.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.shared.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["types.ts"]
+}