--- a/packages/turf-clusters/tsconfig.testTypes.json
+++ b/packages/turf-clusters/tsconfig.testTypes.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.shared.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["types.ts"]
+}