--- a/packages/turf-hex-grid/tsconfig.testTypes.json
+++ b/packages/turf-hex-grid/tsconfig.testTypes.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.shared.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["types.ts"]
+}