--- a/packages/turf-kinks/tsconfig.testTypes.json
+++ b/packages/turf-kinks/tsconfig.testTypes.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.shared.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["types.ts"]
+}