--- a/packages/turf-line-arc/tsconfig.testTypes.json
+++ b/packages/turf-line-arc/tsconfig.testTypes.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.shared.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["types.ts"]
+}