--- a/packages/turf-line-intersect/index.ts
+++ b/packages/turf-line-intersect/index.ts
@@ -9,7 +9,7 @@ import {
   Polygon,
 } from "geojson";
 import type { Intersection } from "sweepline-intersections";
-import { sweeplineIntersections as findIntersections } from "./lib/sweepline-intersections-export.js";
+import sweeplineIntersections from "sweepline-intersections";
 
 /**
  * Takes any LineString or Polygon GeoJSON and returns the intersecting point(s).
@@ -66,7 +66,7 @@ function lineIntersect<
     features.push(feature(line2));
   }
 
-  const intersections = findIntersections(
+  const intersections = sweeplineIntersections(
     featureCollection(features),
     ignoreSelfIntersections
   );