--- a/packages/turf-nearest-point-to-line/tsconfig.testTypes.json
+++ b/packages/turf-nearest-point-to-line/tsconfig.testTypes.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.shared.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["types.ts"]
+}