--- a/packages/turf-nearest-point/tsconfig.testTypes.json
+++ b/packages/turf-nearest-point/tsconfig.testTypes.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.shared.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["types.ts"]
+}