--- a/packages/turf-polygon-smooth/tsconfig.testTypes.json
+++ b/packages/turf-polygon-smooth/tsconfig.testTypes.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.shared.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["types.ts"]
+}