--- a/packages/turf-polygonize/tsconfig.testTypes.json
+++ b/packages/turf-polygonize/tsconfig.testTypes.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.shared.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["types.ts"]
+}