@mfedderly 2024-08-07
Wow this is actually insanely fast

timings on an M1 pro

```
# pnpm lerna run test --scope=@turf/difference
# master / no nx cache
12.04s user 1.22s system 175% cpu 7.545 total
# master / cache everything except test.ts file
1.93s user 0.40s system 87% cpu 2.675 total
# master / cache everything including tests
0.85s user 0.19s system 99% cpu 1.043 total

# this pr / no caching
pnpm test  0.85s user 0.14s system 102% cpu 0.969 total
```