@mfedderly 2024-08-08
I spent some time trying to get typescript project references to work, but I think there's some weird issue with rollup/tsup and project references that I don't fully understand.